pyproject.toml
[project]
name = "vidstr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "google-genai",
    "python-dotenv",
]
